I can't focus in this fucker anymore.
I've tried for a week to work on this, and somehow I feel as though I've
gone backwards. I don't think this game will ever be finished, and if it
ever is, I don't think I'm the right guy to do it. I don't know if it's a
problem with the game, with me or both, but for now, I'm just going to
push this to GitHub, and try to forget about it.